:construction_worker: 배포를 위한 패키지 이름 변경 및 리팩토링
