+++ app/src/main/java/com/umc/ttatta/app/MainApplication.kt
@@ -1,0 +1,20 @@
+package com.umc.ttatta.app
+
+import android.app.Application
+import android.util.Log
+import com.kakao.sdk.common.KakaoSdk
+import com.naver.maps.map.NaverMapSdk
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class MainApplication: Application() {
+    override fun onCreate() {
+        super.onCreate()
+
+        NaverMapSdk.getInstance(this).client =
+            NaverMapSdk.NcpKeyClient(BuildConfig.NAVER_SDK_CLIENT_ID)
+
+        KakaoSdk.init(this, BuildConfig.KAKAO_NATIVE_APP_KEY)
+        Log.d("MainApplication", "Kakao SDK hash Key: ${KakaoSdk.keyHash}")
+    }
+}
