+++ app/src/main/java/com/umc/ttatta/app/model/event/IntentEvent.kt
@@ -1,0 +1,8 @@
+package com.umc.ttatta.app.model.event
+
+import com.umc.ttatta.app.intent.IntentType
+
+data class IntentEvent(
+    val intentType: IntentType,
+    val onDismissed: () -> Unit,
+)
