+++ app/src/main/java/com/umc/ttatta/app/model/prop/CenterButtonProp.kt
@@ -1,0 +1,6 @@
+package com.umc.ttatta.app.model.prop
+
+data class CenterButtonProp(
+    val recordOptionPickerProp: RecordOptionPickerProp,
+    val onDismissed: () -> Unit,
+)
