+++ app/src/main/java/com/umc/ttatta/app/model/prop/NavigationBarProp.kt
@@ -1,0 +1,9 @@
+package com.umc.ttatta.app.model.prop
+
+import com.umc.ttatta.app.component.NavigationItem
+
+data class NavigationBarProp(
+    val currentNavigationItem: NavigationItem?,
+    val onNavigate: (NavigationItem) -> Unit,
+    val onCenterButtonClicked: () -> Unit,
+)
