+++ app/src/main/java/com/umc/ttatta/app/model/prop/RecordOptionPickerProp.kt
@@ -1,0 +1,10 @@
+package com.umc.ttatta.app.model.prop
+
+import com.umc.design.character.AccessorySet
+
+data class RecordOptionPickerProp(
+    val userName: String,
+    val accessories: AccessorySet,
+    val onCameraOptionClicked: () -> Unit,
+    val onGalleryOptionClicked: () -> Unit,
+)
