+++ app/src/main/java/com/umc/ttatta/app/tracking/LocationTrackingService.kt
@@ -1,0 +1,169 @@
+package com.umc.ttatta.app.tracking
+
+import android.Manifest
+import android.app.NotificationChannel
+import android.app.NotificationManager
+import android.app.Service
+import android.content.Context
+import android.content.Intent
+import android.content.pm.PackageManager
+import android.os.Build
+import android.os.IBinder
+import android.os.Looper
+import android.util.Log
+import androidx.annotation.RequiresPermission
+import androidx.core.app.NotificationCompat
+import androidx.core.content.ContextCompat
+import com.google.android.gms.location.FusedLocationProviderClient
+import com.google.android.gms.location.LocationCallback
+import com.google.android.gms.location.LocationRequest
+import com.google.android.gms.location.LocationResult
+import com.google.android.gms.location.LocationServices
+import com.google.android.gms.location.Priority
+import com.umc.core.repository.SettingRepository
+import com.umc.ttatta.app.R
+import com.umc.ttatta.app.util.PermissionManager.checkPermission
+import dagger.hilt.android.AndroidEntryPoint
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
+import javax.inject.Inject
+
+@AndroidEntryPoint
+class LocationTrackingService : Service() {
+
+    @Inject
+    lateinit var settingRepository: SettingRepository
+
+    private lateinit var fusedLocationClient: FusedLocationProviderClient
+    private lateinit var locationCallback: LocationCallback
+
+    companion object {
+        private const val ACTION_START = "ACTION_START_LOCATION_TRACKING"
+        private const val ACTION_STOP = "ACTION_STOP_LOCATION_TRACKING"
+
+        fun Context.startLocationTrackingService() {
+            if (!checkPermission(
+                    Manifest.permission.ACCESS_FINE_LOCATION,
+                    Manifest.permission.ACCESS_COARSE_LOCATION,
+                    Manifest.permission.FOREGROUND_SERVICE_LOCATION,
+                    Manifest.permission.POST_NOTIFICATIONS,
+                )
+            ) {
+                throw Exception("Location permission is not granted")
+            }
+
+            val intent = Intent(this, LocationTrackingService::class.java).apply {
+                action = ACTION_START
+            }
+
+            startService(intent)
+        }
+
+        fun Context.finishLocationTrackingService() {
+            val intent = Intent(this, LocationTrackingService::class.java).apply {
+                action = ACTION_STOP
+            }
+
+            startService(intent)
+        }
+    }
+
+    override fun onCreate() {
+        super.onCreate()
+        fusedLocationClient = LocationServices.getFusedLocationProviderClient(this)
+    }
+
+    @RequiresPermission(
+        allOf = [
+            Manifest.permission.ACCESS_FINE_LOCATION,
+            Manifest.permission.ACCESS_COARSE_LOCATION,
+            Manifest.permission.FOREGROUND_SERVICE_LOCATION,
+            Manifest.permission.POST_NOTIFICATIONS,
+        ]
+    )
+    override fun onStartCommand(intent: Intent?, flags: Int, startId: Int): Int {
+        when (intent?.action) {
+            ACTION_START -> {
+                startForegroundService()
+                startLocationUpdates()
+            }
+
+            ACTION_STOP -> {
+                stopLocationUpdates()
+                stopForeground(STOP_FOREGROUND_REMOVE)
+                stopSelf()
+            }
+        }
+
+        return START_STICKY
+    }
+
+    @RequiresPermission(allOf = [Manifest.permission.ACCESS_FINE_LOCATION, Manifest.permission.ACCESS_COARSE_LOCATION])
+    private fun startLocationUpdates() {
+        val locationRequest = LocationRequest
+            .Builder(Priority.PRIORITY_HIGH_ACCURACY, 600000)  // 10 minutes
+            .setWaitForAccurateLocation(false)
+            .setMinUpdateIntervalMillis(300000)  // 5 minutes
+            .setMaxUpdateDelayMillis(1200000) // 20 minutes
+            .build()
+
+        locationCallback = object : LocationCallback() {
+            override fun onLocationResult(locationResult: LocationResult) {
+                val location = locationResult.lastLocation ?: return
+
+                CoroutineScope(Dispatchers.IO).launch {
+                    runCatching {
+                        settingRepository.sendLocation(location.latitude, location.longitude)
+                    }
+                }
+            }
+        }
+
+        fusedLocationClient.requestLocationUpdates(
+            locationRequest,
+            locationCallback,
+            Looper.getMainLooper(),
+        )
+    }
+
+    private fun stopLocationUpdates() {
+        runCatching {
+            fusedLocationClient.removeLocationUpdates(locationCallback)
+        }
+    }
+
+    @RequiresPermission(allOf = [Manifest.permission.POST_NOTIFICATIONS, Manifest.permission.FOREGROUND_SERVICE_LOCATION])
+    private fun startForegroundService() {
+        val channelId = "location_tracking_channel"
+        val channelName = "위치 전송 알림"
+
+        val channel = NotificationChannel(
+            channelId,
+            channelName,
+            NotificationManager.IMPORTANCE_DEFAULT,
+        )
+
+        val manager = getSystemService(NOTIFICATION_SERVICE) as NotificationManager
+        manager.createNotificationChannel(channel)
+
+        val notification = NotificationCompat
+            .Builder(this, channelId)
+            .setContentTitle("따따")
+            .setContentText("위치 정보를 수집하고 있습니다.")
+            .setSmallIcon(R.drawable.ic_notification)
+            .build()
+
+        startForeground(1, notification)
+    }
+
+
+    override fun onBind(intent: Intent?): IBinder? {
+        return null
+    }
+
+    override fun onDestroy() {
+        super.onDestroy()
+        stopLocationUpdates()
+    }
+}
