+++ app/src/main/java/com/umc/ttatta/app/util/setStatusBarTransparent.kt
@@ -1,0 +1,15 @@
+package com.umc.ttatta.app.util
+
+import android.view.WindowManager
+import androidx.activity.ComponentActivity
+import androidx.core.view.WindowCompat
+
+fun ComponentActivity.setStatusBarTransparent() {
+    window.apply {
+        WindowCompat.setDecorFitsSystemWindows(this, false)
+        setFlags(
+            WindowManager.LayoutParams.FLAG_LAYOUT_NO_LIMITS,
+            WindowManager.LayoutParams.FLAG_LAYOUT_NO_LIMITS
+        )
+    }
+}
