--- app/src/main/java/com/umc/ttatta/component/FinishHandler.kt
@@ -1,24 +1,0 @@
-package com.umc.ttatta.component
-
-import android.app.Activity
-import android.widget.Toast
-import androidx.activity.compose.BackHandler
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableLongStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.platform.LocalContext
-
-@Composable
-fun FinishHandler() {
-    val context = LocalContext.current as Activity
-    var backPressedTime by remember { mutableLongStateOf(0L) }
-
-    BackHandler {
-        val currentTime = System.currentTimeMillis()
-        if (currentTime - backPressedTime < 2000L) context.finish()
-        else Toast.makeText(context, "뒤로 버튼을 한 번 더 누르면 종료됩니다.", Toast.LENGTH_SHORT).show()
-        backPressedTime = currentTime
-    }
-}
