--- app/src/main/java/com/umc/ttatta/model/event/IntentEvent.kt
@@ -1,8 +1,0 @@
-package com.umc.ttatta.model.event
-
-import com.umc.ttatta.intent.IntentType
-
-data class IntentEvent(
-    val intentType: IntentType,
-    val onDismissed: () -> Unit,
-)
