--- app/src/main/java/com/umc/ttatta/model/prop/CenterButtonProp.kt
@@ -1,8 +1,0 @@
-package com.umc.ttatta.model.prop
-
-import com.umc.ttatta.model.prop.RecordOptionPickerProp
-
-data class CenterButtonProp(
-    val recordOptionPickerProp: RecordOptionPickerProp,
-    val onDismissed: () -> Unit,
-)
