--- app/src/main/java/com/umc/ttatta/model/prop/NavigationBarProp.kt
@@ -1,9 +1,0 @@
-package com.umc.ttatta.model.prop
-
-import com.umc.ttatta.component.NavigationItem
-
-data class NavigationBarProp(
-    val currentNavigationItem: NavigationItem?,
-    val onNavigate: (NavigationItem) -> Unit,
-    val onCenterButtonClicked: () -> Unit,
-)
