--- app/src/main/java/com/umc/ttatta/model/routing/CategoryRoutingInfo.kt
@@ -1,5 +1,0 @@
-package com.umc.ttatta.model.routing
-
-class CategoryRoutingInfo(
-    val showTopBar: Boolean,
-)
