--- app/src/main/java/com/umc/ttatta/model/routing/ChallengeRoutingInfo.kt
@@ -1,6 +1,0 @@
-package com.umc.ttatta.model.routing
-
-class ChallengeRoutingInfo(
-    val isPointGranted: Boolean,
-    val isPoppedFromRecord: Boolean,
-)
