--- app/src/main/java/com/umc/ttatta/model/routing/RecordEntryInfo.kt
@@ -1,11 +1,0 @@
-package com.umc.ttatta.model.routing
-
-class RecordEntryInfo(
-    val mode: RecordRoutingOption,
-    val challengeId: Long?,
-) {
-    enum class RecordRoutingOption {
-        CAMERA,
-        GALLERY,
-    }
-}
