--- app/src/main/java/com/umc/ttatta/model/routing/RecordRoutingInfo.kt
@@ -1,8 +1,0 @@
-package com.umc.ttatta.model.routing
-
-import java.io.File
-
-class RecordRoutingInfo(
-    val image: File,
-    val challengeId: Long?,
-)
