--- app/src/main/java/com/umc/ttatta/navigation/NavigationRoute.kt
@@ -1,13 +1,0 @@
-package com.umc.ttatta.navigation
-
-enum class NavigationRoute {
-    SPLASH,
-    LOCK,
-    LOGIN,
-    HOME,
-    FOOTPRINT,
-    CHALLENGE,
-    MY_PAGE,
-    CATEGORY,
-    RECORD,
-}
