--- app/src/main/java/com/umc/ttatta/util/runWithScope.kt
@@ -1,7 +1,0 @@
-package com.umc.ttatta.util
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
-import kotlinx.coroutines.launch
-
-fun <VM: ViewModel, T> VM.runWithScope(block: suspend VM.() -> T) = this.viewModelScope.launch { block() }
